Support Cachix writes via a `cachix push` on each command.
To use this feature, set the CACHIX_NAME environment variable in the
pipeline, where the value of the variable is the name of the cache to
push to.

Note that this feature also requires either that you set the
CACHIX_AUTH_TOKEN pipeline secret for the cache, or that you configure
an auth token in the Buildkite agent's home directory via `cachix
authtoken`.